requirements.txt
# Core
tensorflow==2.13.0
numpy>=1.24.3
pandas>=2.0.3
matplotlib>=3.7.2
scikit-learn>=1.3.0
Pillow>=10.0.0
tqdm>=4.65.0

# Image Processing
opencv-python-headless>=4.8.0

# Streamlit
streamlit>=1.25.0
altair>=5.0.0

# AWS
boto3>=1.28.0
sagemaker>=2.173.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
jupyter>=1.0.0
